UI: Make generator options toggle buttons to make it more clear which are enabled.
