+++ src/planner/menu/component/generator/MenuComponentSymmetry.java
@@ -1,41 +1,10 @@
 package planner.menu.component.generator;
 import multiblock.symmetry.Symmetry;
-import org.lwjgl.glfw.GLFW;
-import planner.Core;
-import simplelibrary.font.FontManager;
-import simplelibrary.opengl.gui.components.MenuComponent;
-public class MenuComponentSymmetry extends MenuComponent{
+import planner.menu.component.MenuComponentToggleButton;
+public class MenuComponentSymmetry extends MenuComponentToggleButton {
     public final Symmetry symmetry;
-    public boolean enabled = false;
     public MenuComponentSymmetry(Symmetry symmetry){
-        super(0, 0, 0, 32);
+        super(0, 0, 0, 32, symmetry.name, symmetry.defaultEnabled());
         this.symmetry = symmetry;
-        enabled = symmetry.defaultEnabled();
-    }
-    @Override
-    public void render(){
-        if(isSelected){
-            if(isMouseOver)Core.applyColor(Core.theme.getMouseoverSelectedComponentColor(Core.getThemeIndex(this)));
-            else Core.applyColor(Core.theme.getSelectedComponentColor(Core.getThemeIndex(this)));
-        }else{
-            if(isMouseOver)Core.applyColor(Core.theme.getMouseoverComponentColor(Core.getThemeIndex(this)));
-            else Core.applyColor(Core.theme.getComponentColor(Core.getThemeIndex(this)));
-        }
-        drawRect(x, y, x+width, y+height, 0);
-        Core.applyColor(Core.theme.getComponentTextColor(Core.getThemeIndex(this)));
-        drawText();
-    }
-    public void drawText(){
-        double textLength = FontManager.getLengthForStringWithHeight(symmetry.name, height);
-        double scale = Math.min(1, width/textLength);
-        double textHeight = (int)(height*scale)-1;
-        drawCenteredText(x, y+height/2-textHeight/2, x+width, y+height/2+textHeight/2, symmetry.name);
-    }
-    @Override
-    public void onMouseButton(double x, double y, int button, boolean pressed, int mods){
-        super.onMouseButton(x, y, button, pressed, mods);
-        if(button==GLFW.GLFW_MOUSE_BUTTON_LEFT&&pressed){
-            enabled = !enabled;
-        }
     }
 }
